backend/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.6.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.2
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==3.4.7
Cython==0.29.23
defusedxml==0.7.1
Deprecated==1.2.12
Django==3.2.5
django-braces==1.14.0
django-cors-headers==3.7.0
django-oauth-toolkit==1.5.0
django-rest-framework-social-oauth2==1.1.0
django-sslserver==0.22
djangorestframework==3.12.4
google-api-core==1.31.0
google-api-python-client==2.14.0
google-auth==1.33.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
ipywidgets==7.6.3
JPype1==1.3.0
jupyter==1.0.0
jupyter-console==6.4.0
jupyterlab-widgets==1.0.0
jwcrypto==0.9.1
konlpy==0.5.2
lxml==4.6.3
numpy==1.21.1
oauthlib==3.1.1
packaging==21.0
pdfminer==20191125
Pillow==9.0.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.1.0
pyparsing==2.4.7
PySocks==1.7.1
python3-openid==3.2.0
pytz==2021.1
qtconsole==5.0.3
QtPy==1.9.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
tweepy==3.10.0
uritemplate==3.0.1
urllib3==1.26.6
widgetsnbextension==3.5.1
wrapt==1.12.1
